Extract data from MSnbase objects for use in Spectra

extractSpectraData extracts the spectra data (m/z and intensity values including metadata) from MSnbase::Spectrum1 or MSnbase::Spectrum2 objects (or list of such objects) and returns these as a DataFrame that can be used to create a Spectra::Spectra object. This function enables thus to convert data from the old MSnbase package to the newer Spectra package.

extractSpectraData(x)

Arguments

x
a list of MSnbase::Spectrum objects or an object extending MSnbase::MSnExp (such as also the xcms::XCMSnExp object).

Value

S4Vectors::DataFrame with the full spectrum data that can be passed along to the Spectra::Spectra() function to create such an object.

Author

Johannes Rainer

Examples

## Read an mzML file with MSnbase
fl <- system.file("TripleTOF-SWATH", "PestMix1_SWATH.mzML",
    package = "msdata")
data <- MSnbase::filterRt(MSnbase::readMSData(fl, mode = "onDisk"),
    rt = c(1, 6))

## Extract the data as a DataFrame
res <- extractSpectraData(data)

## This can be used as an input for the Spectra constructor of the
## Spectra package.
sps <- Spectra::Spectra(res)
sps

#> MSn data (Spectra) with 50 spectra in a MsBackendDataFrame backend:
#>            msLevel     rtime scanIndex
#>          <integer> <numeric> <integer>
#> F1.S0010         2        NA        NA
#> F1.S0011         2        NA        NA
#> F1.S0012         2        NA        NA
#> F1.S0013         2        NA        NA
#> F1.S0014         2        NA        NA
#> ...            ...       ...       ...
#> F1.S0055         2        NA        NA
#> F1.S0056         2        NA        NA
#> F1.S0057         2        NA        NA
#> F1.S0058         2        NA        NA
#> F1.S0059         2        NA        NA
#>  ... 43 more variables/columns.
#> Processing:
#>