Grouping of LC-MS features

After correspondence analysis (xcms::groupChromPeaks()) the identified LC-MS features can be further grouped based on different criteria into feature groups which contain ideally features representing ions (adducts, isotopes) of the same compound/metabolite.

See MsFeatures::groupFeatures() for an overview on the general feature grouping concept as well as details on the individual settings and parameters.

The available options for groupFeatures on xcms preprocessing results (i.e. on XCMSnExp objects after correspondence analysis with groupChromPeaks()) are:

Grouping by similar retention times: groupFeatures-similar-rtime().

Grouping by similar feature values across samples: AbundanceSimilarityParam().

Grouping by similar peak shape of extracted ion chromatograms: EicCorrelationParam().

An ideal workflow grouping features should sequentially perform the above methods (in the listed order).

Feature groups can be accessed with the featureGroups function.

# S4 method for XCMSnExp
featureGroups(object)

# S4 method for XCMSnExp
featureGroups(object) <- value

Arguments

object
an XCMSnExp() object.

value
for featureGroups<-: replacement for the feature groups in object. Has to be of length 1 or length equal to the number of features in object.

See also

Author

Johannes Rainer, Mar Garcia-Aloy, Vinicius Veri Hernandes