Group EICs based on their correlation

groupEicCorrelation groups (extracted ion) chromatograms (EICs) based on their correlation with each other. If this correlation is >= than the provided threshold they are grouped.

If x is a Chromatograms() object with more than one column (sample), pairwise correlations between EICs are first calculated for each column (sample) of x separately and subsequently aggregated across samples using aggregationFun. If x is a Chromatograms with 4 rows (EICs) and 3 columns (samples), pairwise correlations are first calculated between all 4 EICs in each of the 3 columns resulting in 3 correlation matrices (of dimension 4x4). These correlation matrices are combined into a single matrix by combining the 3 correlation values per comparison with aggregationFun. By default the mean of the correlation value between e.g. EIC 1 and EIC 2 in each of the 3 columns is used as the final correlation value. Similar to the one-column case EICs are grouped if their (aggregated) correlation coefficient is larger than threshold.

Two types of groupings are available:

greedy = FALSE (the default): the algorithm creates small groups of highly correlated members, all of which have a correlation with each other that are >= threshold. Note that with this algorithm, rows in x could still have a correlation >= threshold with one or more elements of a group they are not part of. See notes below for more information.

greedy = TRUE: the algorithm creates large groups containing rows that have a correlation >= threshold with at least one element of that group. For example, if row 1 and 3 have a correlation above the threshold and rows 3 and 5 too (but correlation between 1 and 5 is below the threshold) all 3 are grouped into the same group (i.e. rows 1, 3 and 5).

For more information see groupByCorrelation().

groupEicCorrelation(
  x,
  aggregationFun = mean,
  threshold = 0.8,
  greedy = FALSE,
  ...
)

Arguments

x
Chromatograms() object of list of Chromatogram() objects.

aggregationFun
function to combine the correlation values between pairs of EICs across samples (columns). See description for details.

threshold
numeric(1) with the threshold for correlation above which EICs are grouped together.

greedy
logical(1) whether a version of the grouping algorithm should be used that leads to larger, more loosely correlated, groups. The default is greedy = FALSE. See description for more information.

...
parameters for the correlate() function for Chromatograms() objects.

Value

factor same length as nrow(x) (if x is a Chromatograms object) or length(x) (if x is a list) with the group each EIC is assigned to.

See also

Author

Johannes Rainer

Examples

library(MSnbase)
set.seed(123)
chr1 <- Chromatogram(rtime = 1:10 + rnorm(n = 10, sd = 0.3),
    intensity = c(5, 29, 50, NA, 100, 12, 3, 4, 1, 3))
chr2 <- Chromatogram(rtime = 1:10 + rnorm(n = 10, sd = 0.3),
    intensity = c(80, 50, 20, 10, 9, 4, 3, 4, 1, 3))
chr3 <- Chromatogram(rtime = 3:9 + rnorm(7, sd = 0.3),
    intensity = c(53, 80, 130, 15, 5, 3, 2))
chrs <- Chromatograms(list(chr1, chr2, chr3))

groupEicCorrelation(chrs)

#> [1] 1 2 1
#> Levels: 1 2

## With a Chromatograms with two columns, use the maximal correlation
## coefficient found in each of the columns
chrs <- Chromatograms(list(chr1, chr2, chr3, chr1, chr2, chr3), ncol = 2)
groupEicCorrelation(chrs, aggregationFun = max)

#> [1] 1 2 1
#> Levels: 1 2