Group features based on correlation of abundances across samples

This methods groups features based on correlation of abundances (i.e. feature values) across samples. Parameter subset allows to define a sub set of samples on which this correlation should be correlated. It might for example be better to exclude QC samples from this correlation analysis because values of all features are supposed to be constant in these samples and including these could bias the correlation estimation. Also, it might be better to perform the correlation including also gap filled values which is possible with parameter filled = TRUE on an object on which also xcms::fillChromPeaks() was called.

AbundanceCorrelationParam(
  threshold = 0.9,
  value = "into",
  method = c("maxint", "medret", "sum"),
  intensity = "maxo",
  filled = TRUE,
  subset = integer(),
  greedy = FALSE,
  transform = log2
)

# S4 method for XCMSnExp,AbundanceCorrelationParam
groupFeatures(object, param, msLevel = 1L)

Arguments

threshold
numeric(1) with the minimal required correlation coefficient to group featues.

value
character(1) passed to the featureValues call. See featureValues() for details. Defaults to value = "into".

method
character(1) passed to the featureValues call. See featureValues() for details. Defaults to method = "maxint".

intensity
character(1) passed to the featureValues call. See featureValues() for details. Defaults to intensity = "maxo".

filled
logical(1) whether filled-in values should be included in the correlation analysis. Defaults to filled = TRUE.

subset
integer or logical defining a subset of samples (at least 2) on which the correlation should be performed.

greedy
logical(1) which grouping algorithm should be used: one that creates small groups of highly correlated features (greedy = FALSE, the default) or whether features should be grouped that have at least one correlation with any other member of the group in common (greedy = TRUE). See groupByCorrelation() for details.

transform
function to be applied to the feature intensities prior correlation. Defaults to transform = log2, i.e. log2 transforms the data before correlation. To use the values as-is use transform = identity.

object
XCMSnExp() object containing also correspondence results.

param
AbudanceCorrelationParam object with the settings for the method.

msLevel
integer(1) defining the MS level on which the features should be grouped.

Value

input XCMSnExp with feature groups added (i.e. in column "feature_group" of its featureDefinitions data frame.

See also

Examples

## Performing a quick preprocessing of a test data set.
library(faahKO)

#> Loading required package: xcms

#> Loading required package: BiocParallel

#> 
#> This is xcms version 3.11.1 

#> 
#> Attaching package: ‘xcms’

#> The following object is masked from ‘package:stats’:
#> 
#>     sigma

fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

#> Processing 2867 mz slices ... 

#> OK

## Group features based on correlation of feature values (integrated
## peak area) across samples. Note that there are many missing values
## in the feature value which influence grouping of features in the present
## data set.
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8))
table(featureDefinitions(xodg_grp)$feature_group)

#> 
#>   FG.1  FG.10 FG.100 FG.101 FG.102 FG.103 FG.104 FG.105 FG.106 FG.107 FG.108 
#>      2      1      1      1      1      1      1      1      1      1      1 
#> FG.109  FG.11 FG.110 FG.111 FG.112 FG.113 FG.114 FG.115 FG.116 FG.117 FG.118 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.119  FG.12 FG.120 FG.121 FG.122 FG.123  FG.13  FG.14  FG.15  FG.16  FG.17 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.18  FG.19   FG.2  FG.20  FG.21  FG.22  FG.23  FG.24  FG.25  FG.26  FG.27 
#>      1      1      8      1      1      1      1      1      1      1      1 
#>  FG.28  FG.29   FG.3  FG.30  FG.31  FG.32  FG.33  FG.34  FG.35  FG.36  FG.37 
#>      1      1     15      1      1      1      1      1      1      1      1 
#>  FG.38  FG.39   FG.4  FG.40  FG.41  FG.42  FG.43  FG.44  FG.45  FG.46  FG.47 
#>      1      1      4      1      1      1      1      1      1      1      1 
#>  FG.48  FG.49   FG.5  FG.50  FG.51  FG.52  FG.53  FG.54  FG.55  FG.56  FG.57 
#>      1      1      4      1      1      1      1      1      1      1      1 
#>  FG.58  FG.59   FG.6  FG.60  FG.61  FG.62  FG.63  FG.64  FG.65  FG.66  FG.67 
#>      1      1      6      1      1      1      1      1      1      1      1 
#>  FG.68  FG.69   FG.7  FG.70  FG.71  FG.72  FG.73  FG.74  FG.75  FG.76  FG.77 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.78  FG.79   FG.8  FG.80  FG.81  FG.82  FG.83  FG.84  FG.85  FG.86  FG.87 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.88  FG.89   FG.9  FG.90  FG.91  FG.92  FG.93  FG.94  FG.95  FG.96  FG.97 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.98  FG.99 
#>      1      1 

## Group based on the maximal peak intensity per feature
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8, value = "maxo"))
table(featureDefinitions(xodg_grp)$feature_group)

#> 
#>   FG.1  FG.10 FG.100 FG.101 FG.102 FG.103 FG.104 FG.105 FG.106 FG.107 FG.108 
#>      6      1      1      1      1      1      1      1      1      1      1 
#> FG.109  FG.11 FG.110 FG.111 FG.112 FG.113 FG.114 FG.115 FG.116 FG.117 FG.118 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.119  FG.12 FG.120 FG.121 FG.122  FG.13  FG.14  FG.15  FG.16  FG.17  FG.18 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.19   FG.2  FG.20  FG.21  FG.22  FG.23  FG.24  FG.25  FG.26  FG.27  FG.28 
#>      1      4      1      1      1      1      1      1      1      1      1 
#>  FG.29   FG.3  FG.30  FG.31  FG.32  FG.33  FG.34  FG.35  FG.36  FG.37  FG.38 
#>      1     22      1      1      1      1      1      1      1      1      1 
#>  FG.39   FG.4  FG.40  FG.41  FG.42  FG.43  FG.44  FG.45  FG.46  FG.47  FG.48 
#>      1      5      1      1      1      1      1      1      1      1      1 
#>  FG.49   FG.5  FG.50  FG.51  FG.52  FG.53  FG.54  FG.55  FG.56  FG.57  FG.58 
#>      1      2      1      1      1      1      1      1      1      1      1 
#>  FG.59   FG.6  FG.60  FG.61  FG.62  FG.63  FG.64  FG.65  FG.66  FG.67  FG.68 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.69   FG.7  FG.70  FG.71  FG.72  FG.73  FG.74  FG.75  FG.76  FG.77  FG.78 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.79   FG.8  FG.80  FG.81  FG.82  FG.83  FG.84  FG.85  FG.86  FG.87  FG.88 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.89   FG.9  FG.90  FG.91  FG.92  FG.93  FG.94  FG.95  FG.96  FG.97  FG.98 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.99 
#>      1