Overlay plots from multiple EICs

plotOverlay draws chromatographic peak data from multiple (different) extracted ion chromatograms (EICs) into the same plot. This allows to directly compare the peak shape of these EICs in the same sample. In contrast to the plot function for Chromatograms() object, which draws the data from the same EIC across multiple samples in the same plot, this function draws the different EICs from the same sample into the same plot.

plotOverlay(
  x,
  col = "#00000060",
  type = "l",
  main = NULL,
  xlab = "rtime",
  ylab = "intensity",
  ...
)

Arguments

x
Chromatograms() object with the EICs to draw.

col
color representation. Should be either of length 1 (to use the same color for each EIC) or equal to nrow(x) to use a different color for each EIC (row in x).

type
character(1) specifying the plot type (see help from base plot function). type = "l" (the default) draws lines, type = "p" would draw each intensity as a point etc.

main
optional character with the title for the plot of each sample. Can be either of length 1 or equal ncol(x).

xlab
character(1) with the label for the x-axis.

ylab
character(1) with the label for the y-axis.

...
additional parameters to be passed down to the points function.

Author

Johannes Rainer