Group features based on correlation of abundances across samples

This methods groups features based on correlation of abundances (i.e. feature values) across samples. Parameter subset allows to define a sub set of samples on which this correlation should be correlated. It might for example be better to exclude QC samples from this correlation analysis because values of all features are supposed to be constant in these samples and including these could bias the correlation estimation. Also, it might be better to perform the correlation including also gap filled values which is possible with parameter filled = TRUE on an object on which also xcms::fillChromPeaks() was called.

AbundanceCorrelationParam(
  threshold = 0.9,
  value = "into",
  method = c("maxint", "medret", "sum"),
  intensity = "maxo",
  filled = TRUE,
  subset = integer(),
  greedy = FALSE,
  transform = log2
)

# S4 method for XCMSnExp,AbundanceCorrelationParam
groupFeatures(object, param, msLevel = 1L)

Arguments

threshold
numeric(1) with the minimal required correlation coefficient to group featues.

value
character(1) passed to the featureValues call. See featureValues() for details. Defaults to value = "into".

method
character(1) passed to the featureValues call. See featureValues() for details. Defaults to method = "maxint".

intensity
character(1) passed to the featureValues call. See featureValues() for details. Defaults to intensity = "maxo".

filled
logical(1) whether filled-in values should be included in the correlation analysis. Defaults to filled = TRUE.

subset
integer or logical defining a subset of samples (at least 2) on which the correlation should be performed.

greedy
logical(1) which grouping algorithm should be used: one that creates small groups of highly correlated features (greedy = FALSE, the default) or whether features should be grouped that have at least one correlation with any other member of the group in common (greedy = TRUE). See groupByCorrelation() for details.

transform
function to be applied to the feature intensities prior correlation. Defaults to transform = log2, i.e. log2 transforms the data before correlation. To use the values as-is use transform = identity.

object
XCMSnExp() object containing also correspondence results.

param
AbudanceCorrelationParam object with the settings for the method.

msLevel
integer(1) defining the MS level on which the features should be grouped.

Value

input XCMSnExp with feature groups added (i.e. in column "feature_group" of its featureDefinitions data frame.

See also

Author

Johannes Rainer

Examples

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 937 regions of interest ...

#>  OK: 87 found.

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 1025 regions of interest ...

#>  OK: 100 found.

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 623 regions of interest ...

#>  OK: 71 found.

pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

#> Processing 2867 mz slices ... 

#> OK

## Group features based on correlation of feature values (integrated
## peak area) across samples. Note that there are many missing values
## in the feature value which influence grouping of features in the present
## data set.
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8))
table(featureDefinitions(xodg_grp)$feature_group)

#> 
#> FG.001 FG.002 FG.003 FG.004 FG.005 FG.006 FG.007 FG.008 FG.009 FG.010 FG.011 
#>      2      8     15      4      4      6      1      1      1      1      1 
#> FG.012 FG.013 FG.014 FG.015 FG.016 FG.017 FG.018 FG.019 FG.020 FG.021 FG.022 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.023 FG.024 FG.025 FG.026 FG.027 FG.028 FG.029 FG.030 FG.031 FG.032 FG.033 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.034 FG.035 FG.036 FG.037 FG.038 FG.039 FG.040 FG.041 FG.042 FG.043 FG.044 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.045 FG.046 FG.047 FG.048 FG.049 FG.050 FG.051 FG.052 FG.053 FG.054 FG.055 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.056 FG.057 FG.058 FG.059 FG.060 FG.061 FG.062 FG.063 FG.064 FG.065 FG.066 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.067 FG.068 FG.069 FG.070 FG.071 FG.072 FG.073 FG.074 FG.075 FG.076 FG.077 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.078 FG.079 FG.080 FG.081 FG.082 FG.083 FG.084 FG.085 FG.086 FG.087 FG.088 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.089 FG.090 FG.091 FG.092 FG.093 FG.094 FG.095 FG.096 FG.097 FG.098 FG.099 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.100 FG.101 FG.102 FG.103 FG.104 FG.105 FG.106 FG.107 FG.108 FG.109 FG.110 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.111 FG.112 FG.113 FG.114 FG.115 FG.116 FG.117 FG.118 FG.119 FG.120 FG.121 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.122 FG.123 
#>      1      1 

## Group based on the maximal peak intensity per feature
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8, value = "maxo"))
table(featureDefinitions(xodg_grp)$feature_group)

#> 
#> FG.001 FG.002 FG.003 FG.004 FG.005 FG.006 FG.007 FG.008 FG.009 FG.010 FG.011 
#>      6      4     22      5      2      1      1      1      1      1      1 
#> FG.012 FG.013 FG.014 FG.015 FG.016 FG.017 FG.018 FG.019 FG.020 FG.021 FG.022 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.023 FG.024 FG.025 FG.026 FG.027 FG.028 FG.029 FG.030 FG.031 FG.032 FG.033 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.034 FG.035 FG.036 FG.037 FG.038 FG.039 FG.040 FG.041 FG.042 FG.043 FG.044 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.045 FG.046 FG.047 FG.048 FG.049 FG.050 FG.051 FG.052 FG.053 FG.054 FG.055 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.056 FG.057 FG.058 FG.059 FG.060 FG.061 FG.062 FG.063 FG.064 FG.065 FG.066 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.067 FG.068 FG.069 FG.070 FG.071 FG.072 FG.073 FG.074 FG.075 FG.076 FG.077 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.078 FG.079 FG.080 FG.081 FG.082 FG.083 FG.084 FG.085 FG.086 FG.087 FG.088 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.089 FG.090 FG.091 FG.092 FG.093 FG.094 FG.095 FG.096 FG.097 FG.098 FG.099 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.100 FG.101 FG.102 FG.103 FG.104 FG.105 FG.106 FG.107 FG.108 FG.109 FG.110 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.111 FG.112 FG.113 FG.114 FG.115 FG.116 FG.117 FG.118 FG.119 FG.120 FG.121 
#>      1      1      1      1      1      1      1      1      1      1      1 
#> FG.122 
#>      1