Group features based on correlation of extracted ion chromatograms

Group features based on correlation of their extracted ion chromatograms. This correlation is performed separately for each sample with the correlation coefficients being aggregated across samples for the final comparison with parameter threshold (the 75% quantile of the per-sample correlation values is used for the comparison with threshold).

This feature grouping should be called after an initial feature grouping by retention time (see SimilarRtimeParam()). The feature groups defined in columns "feature_group" of featureDefinitions(object) (for features matching msLevel) will be used and refined by this method. Features with a value of NA in featureDefinitions(object)$feature_group will be skipped/not considered for feature grouping.

While being possible to be performed on the full data set without prior feature grouping , this is not suggested for the following reasons: I) the selection of the top n samples with the highest signal for the feature group will be biased by very abundant compounds as this is performed on the full data set (i.e. the samples with the highest overall intensities are used for correlation of all features) and II) it is computationally much more expensive because a pairwise correlation between all features has to be performed.

It is also suggested to perform the correlation on a subset of samples per feature with the highest intensities of the peaks (for that feature) although it would also be possible to run the correlation on all samples by setting n equal to the total number of samples in the data set. EIC correlation should however be performed ideally on samples in which the original compound is highly abundant to avoid correlation of missing values or noisy peak shapes as much as possible.

By default also the signal which is outside identified chromatographic peaks is excluded from the correlation (parameter clean).

EicCorrelationParam(
  threshold = 0.9,
  n = 1,
  clean = TRUE,
  value = c("maxo", "into"),
  greedy = FALSE
)

# S4 method for XCMSnExp,EicCorrelationParam
groupFeatures(object, param, msLevel = 1L)

Arguments

threshold
numeric(1) with the minimal required correlation coefficient to group featues.

n
numeric(1) defining the total number of samples per feature group on which this correlation should be performed. This value is rounded up to the next larger integer value.

clean
logical(1) whether the correlation should be performed only on the signals within the identified chromatographic peaks (clean = TRUE, default) or all the signal from the extracted ion chromatogram.

value
character(1) defining whether samples should be grouped based on the sum of the maximal peak intensity (value = "maxo", the default) or the integrated peak area (value = "into") for a feature.

greedy
logical(1) which grouping algorithm should be used: one that creates small groups of highly correlated features (greedy = FALSE, the default) or whether features should be grouped that have at least one correlation with any other member of the group in common (greedy = TRUE). See groupByCorrelation() for more information.

object
XCMSnExp() object containing also correspondence results.

param
EicCorrelationParam object with the settings for the method.

msLevel
integer(1) defining the MS level on which the features should be grouped.

Value

input XCMSnExp with feature groups added (i.e. in column "feature_group" of its featureDefinitions data frame.

See also

Author

Johannes Rainer

Examples

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")

#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.

xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 937 regions of interest ...

#>  OK: 87 found.

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 1025 regions of interest ...

#>  OK: 100 found.

#> Detecting mass traces at 25 ppm ... 

#> OK

#> Detecting chromatographic peaks in 623 regions of interest ...

#>  OK: 71 found.

pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

#> Processing 2867 mz slices ... 

#> OK

## Performing a feature grouping based on EIC correlation on a single
## sample
xodg_grp <- groupFeatures(xodg, param = EicCorrelationParam(n = 1))

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

table(featureDefinitions(xodg_grp)$feature_group)

#> 
#>  FG.01  FG.02  FG.03  FG.04  FG.05  FG.06  FG.07  FG.08  FG.09  FG.10 FG.100 
#>      2      2      2      2      3      3      2      2      3      5      1 
#> FG.101 FG.102 FG.103 FG.104 FG.105 FG.106 FG.107 FG.108 FG.109  FG.11 FG.110 
#>      1      1      1      1      1      1      1      1      1      4      1 
#> FG.111 FG.112 FG.113 FG.114 FG.115 FG.116 FG.117 FG.118 FG.119  FG.12 FG.120 
#>      1      1      1      1      1      1      1      1      1      2      1 
#>  FG.13  FG.14  FG.15  FG.16  FG.17  FG.18  FG.19  FG.20  FG.21  FG.22  FG.23 
#>      2      2      2      2      2      2      2      2      2      2      2 
#>  FG.24  FG.25  FG.26  FG.27  FG.28  FG.29  FG.30  FG.31  FG.32  FG.33  FG.34 
#>      2      3      2      2      1      1      1      1      1      1      1 
#>  FG.35  FG.36  FG.37  FG.38  FG.39  FG.40  FG.41  FG.42  FG.43  FG.44  FG.45 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.46  FG.47  FG.48  FG.49  FG.50  FG.51  FG.52  FG.53  FG.54  FG.55  FG.56 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.57  FG.58  FG.59  FG.60  FG.61  FG.62  FG.63  FG.64  FG.65  FG.66  FG.67 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.68  FG.69  FG.70  FG.71  FG.72  FG.73  FG.74  FG.75  FG.76  FG.77  FG.78 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.79  FG.80  FG.81  FG.82  FG.83  FG.84  FG.85  FG.86  FG.87  FG.88  FG.89 
#>      1      1      1      1      1      1      1      1      1      1      1 
#>  FG.90  FG.91  FG.92  FG.93  FG.94  FG.95  FG.96  FG.97  FG.98  FG.99 
#>      1      1      1      1      1      1      1      1      1      1 

## Usually it is better to perform this correlation on pre-grouped features
## e.g. based on approximate retention time.
xodg_grp <- groupFeatures(xodg, param = SimilarRtimeParam(diffRt = 4))
xodg_grp <- groupFeatures(xodg_grp, param = EicCorrelationParam(n = 1))

#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  15%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%

table(featureDefinitions(xodg_grp)$feature_group)

#> 
#> FG.001.1 FG.002.1 FG.003.1 FG.004.1 FG.005.1 FG.006.1 FG.007.1 FG.008.1 
#>        1        3        1        1        1        1        1        1 
#> FG.009.1 FG.009.2 FG.010.1 FG.011.1 FG.012.1 FG.012.2 FG.012.3 FG.012.4 
#>        4        1        1        1        2        1        1        1 
#> FG.013.1 FG.013.2 FG.014.1 FG.015.1 FG.016.1 FG.016.2 FG.017.1 FG.018.1 
#>        1        1        1        1        1        1        1        1 
#> FG.019.1 FG.019.2 FG.019.3 FG.020.1 FG.021.1 FG.022.1 FG.022.2 FG.023.1 
#>        2        1        1        1        2        1        1        1 
#> FG.024.1 FG.025.1 FG.026.1 FG.027.1 FG.028.1 FG.029.1 FG.030.1 FG.031.1 
#>        1        1        1        1        1        1        1        1 
#> FG.032.1 FG.033.1 FG.034.1 FG.035.1 FG.036.1 FG.037.1 FG.037.2 FG.038.1 
#>        1        1        1        1        2        1        1        1 
#> FG.039.1 FG.040.1 FG.041.1 FG.041.2 FG.042.1 FG.043.1 FG.044.1 FG.045.1 
#>        1        1        1        1        2        1        1        2 
#> FG.045.2 FG.046.1 FG.046.2 FG.047.1 FG.048.1 FG.049.1 FG.050.1 FG.051.1 
#>        1        1        1        1        2        1        1        2 
#> FG.052.1 FG.053.1 FG.054.1 FG.055.1 FG.056.1 FG.056.2 FG.057.1 FG.057.2 
#>        1        1        4        2        2        1        2        1 
#> FG.057.3 FG.058.1 FG.059.1 FG.060.1 FG.061.1 FG.062.1 FG.063.1 FG.064.1 
#>        1        1        1        1        1        1        1        1 
#> FG.065.1 FG.066.1 FG.066.2 FG.067.1 FG.067.2 FG.068.1 FG.068.2 FG.069.1 
#>        1        2        1        3        1        1        1        1 
#> FG.069.2 FG.070.1 FG.071.1 FG.071.2 FG.071.3 FG.071.4 FG.071.5 FG.071.6 
#>        1        2        2        1        1        1        1        1 
#> FG.072.1 FG.073.1 FG.074.1 FG.074.2 FG.075.1 FG.076.1 FG.077.1 FG.078.1 
#>        1        1        1        1        1        2        2        1 
#> FG.078.2 FG.079.1 FG.079.2 FG.080.1 FG.081.1 FG.082.1 FG.083.1 FG.084.1 
#>        1        1        1        2        1        1        1        1 
#> FG.084.2 FG.085.1 FG.085.2 FG.086.1 FG.087.1 FG.088.1 FG.089.1 FG.090.1 
#>        1        2        1        1        1        2        1        1 
#> FG.091.1 FG.092.1 FG.092.2 FG.092.3 FG.093.1 FG.094.1 
#>        1        1        1        1        2        1